draft pep 3156 support
In order to support asynchronous usage an OmapiProtocol class is added.
It aims to somewhat support PEP 3156 usage, but still misses a few
details. The implementation is tailored to suit the needs of the Omapi
class to retain backwards compatibility. Methods such as connection_lost
are missing and exceptions are raised where they should not propagate.

git-svn-id: http://pypureomapi.googlecode.com/svn/trunk@68 22607858-6fb5-3b0b-32ac-9b2939425112