Add a new method add_host_supersede
We want to set hostname, routers and domain in the dhcp lease file.
The method add_host_supersede accepts router,domain and hostname
and sets those statements using omapi in the lease file. The method
also handles case when any of the statements are missing/not passed
by the user.